@@ -1,11 +1,16 @@
 module weatherdata
 
-go 1.14
+go 1.19
 
 require (
-	github.com/antchfx/xmlquery v1.3.8
-	github.com/integrii/flaggy v1.4.4
-	github.com/mattn/go-sqlite3 v1.14.9
-	github.com/olekukonko/tablewriter v0.0.5
-	golang.org/x/net v0.0.0-20210805182204-aaa1db679c0d // indirect
+	github.com/antchfx/xmlquery v1.3.15
+	github.com/integrii/flaggy v1.5.2
+	github.com/mattn/go-sqlite3 v1.14.16
+)
+
+require (
+	github.com/antchfx/xpath v1.2.4 // indirect
+	github.com/golang/groupcache v0.0.0-20210331224755-41bb18bfe9da // indirect
+	golang.org/x/net v0.7.0 // indirect
+	golang.org/x/text v0.7.0 // indirect
 )
